[Feat] 보틀 보관함 기능 구현 (#35)
* [feat] : BottleService - 핑퐁 리스트 조회 구현

* [chore] : PingPongListResponse 패키지 경로 수정

* [feat] : GetPingPongListUseCase 구현

* [refactor] : UseCaseModule 분리

* [feat] : KakaoSignInUpReponse 데이터 추가

* [feat] : PingPongRepository 구현

* [refactor] : Bottle의 Id 타입 Long 으로 변경

* [feat] : 보틀 보관함 기능 구현

+++ core/navigator/src/main/kotlin/BottleNavigator.kt
@@ -6,6 +6,6 @@ sealed class BottleNavigator {
     data object ArrivedBottles : MainNavigator()
 
     @Serializable
-    data class Bottle(val bottleId: Int) : MainNavigator()
+    data class Bottle(val bottleId: Long) : MainNavigator()
 
 }
